move op test to own file and test
