+++ rllib/BUILD
@@ -1023,7 +1023,7 @@ py_test(
 # Impala
 py_test(
     name = "test_impala",
-    tags = ["team:rllib", "algorithms_dir", "multi_gpu"],
+    tags = ["team:rllib", "algorithms_dir"],
     size = "large",
     srcs = ["algorithms/impala/tests/test_impala.py"]
 )
@@ -1032,6 +1032,12 @@ py_test(
     tags = ["team:rllib", "algorithms_dir"],
     size = "small",
     srcs = ["algorithms/impala/tests/test_vtrace.py"]
+)
+py_test(
+    name = "test_impala_off_policyness",
+    tags = ["team:rllib", "algorithms_dir", "multi_gpu"],
+    size = "large",
+    srcs = ["algorithms/impala/tests/test_impala_off_policyness.py"]
 )
 
 # MARWIL
