+++ rllib/algorithms/impala/tests/test_impala.py
@@ -8,7 +8,6 @@ from ray.rllib.utils.metrics.learner_info import LEARNER_INFO, LEARNER_STATS_KEY
 from ray.rllib.utils.test_utils import (
     check,
     check_compute_single_action,
-    check_off_policyness,
     check_train_results,
     framework_iterator,
 )
@@ -113,31 +112,6 @@ class TestIMPALA(unittest.TestCase):
                 assert lr3 <= lr2, (lr2, lr3)
                 assert lr3 < lr1, (lr1, lr3)
             finally:
-                algo.stop()
-
-    def test_impala_off_policyness(self):
-        """Test whether Impala can be built with both frameworks."""
-        config = (
-            impala.ImpalaConfig()
-            .environment("CartPole-v1")
-            .resources(num_gpus=1)
-            .rollouts(num_rollout_workers=2)
-        )
-        num_iterations = 4
-
-        for _ in framework_iterator(config, with_eager_tracing=True):
-            for num_aggregation_workers in [0, 1]:
-                config.num_aggregation_workers = num_aggregation_workers
-                print("aggregation-workers={}".format(config.num_aggregation_workers))
-                algo = config.build()
-                for i in range(num_iterations):
-                    results = algo.train()
-                    off_policy_ness = check_off_policyness(results, upper_limit=2.0)
-                    print(f"off-policy'ness={off_policy_ness}")
-
-                check_compute_single_action(
-                    algo,
-                )
                 algo.stop()
